+++ src/services/ui/src/pages/dashboard/index.tsx
@@ -1,12 +1,10 @@
-import { useSearch, SearchData } from "../../api/useSearch";
-import { format } from "date-fns";
 import * as UI from "@enterprise-cmcs/macpro-ux-lib";
-import { LoadingSpinner, ErrorAlert, SearchForm } from "@/components";
-import { ChangeEvent, useEffect, useState } from "react";
-import { Link, redirect } from "react-router-dom";
+import { ChangeEvent, useState } from "react";
+import { redirect } from "react-router-dom";
 import { QueryClient } from "@tanstack/react-query";
 import { getUser } from "@/api/useGetUser";
-import { DataGrid, GridToolbar } from "@mui/x-data-grid";
+import { WaiversList } from "./Lists/waivers";
+import { SpasList } from "./Lists/spas";
 
 const loader = (queryClient: QueryClient) => {
   return async () => {
@@ -31,39 +29,10 @@ export const dashboardLoader = loader;
 
 export const Dashboard = () => {
   const [selectedState, setSelectedState] = useState("VA");
-  const [rowSelectionModel, setRowSelectionModel] = useState<string>();
-  const [searchText, setSearchText] = useState<string>("");
-  const [searchData, setSearchData] = useState<SearchData[] | null>(null);
-  const { mutateAsync, isLoading, error } = useSearch({
-    // Optional: You can provide onSuccess and onError callbacks if needed.
-    // onSuccess: (data) => {},
-    // onError: (error) => {},
-  });
-
-  useEffect(() => {
-    handleSearch(searchText);
-  }, [selectedState]);
-
-  const handleSearch = async (searchText: string) => {
-    try {
-      const data = await mutateAsync({
-        selectedState,
-        searchString: searchText,
-        // programType: "WAIVER",
-        programType: "CHIP OR MEDICAID", // This should be set based on the 'tab' your in
-      });
-
-      setSearchData(data.hits);
-    } catch (error) {
-      console.error("Error occurred during search:", error);
-    }
-  };
 
   const handleStateChange = (event: ChangeEvent<HTMLSelectElement>) => {
     setSelectedState(event.target.value);
   };
-
-  if (isLoading) return <LoadingSpinner />;
 
   return (
     <div className="max-w-screen-lg mx-auto px-4 lg:px-8">
@@ -85,214 +54,22 @@ export const Dashboard = () => {
           </select>
         </div>
       </div>
-      {error ? (
-        <ErrorAlert error={error} />
-      ) : (
-        <div
-          style={{
-            width: "100%",
-            alignItems: "center",
-            justifyContent: "center",
-          }}
-        >
-          <UI.Tabs>
-            <UI.TabPanel id="tab-panel--summary" tabLabel="Waivers">
-              <SearchForm
-                handleSearch={handleSearch}
-                setSearchText={setSearchText}
-                searchText={searchText}
-              />
-              <DataGrid
-                columns={[
-                  {
-                    field: "Transmittal ID Number (TIN)",
-                    hideable: false,
-                    flex: 1,
-                    renderCell(params) {
-                      return (
-                        <Link
-                          className="cursor-pointer text-blue-600"
-                          to={`/record?region=${encodeURIComponent(
-                            params.row._source.seatool.STATE_PLAN.STATE_CODE
-                          )}&id=${encodeURIComponent(params.row._id)}`}
-                        >
-                          {params.row._id}
-                        </Link>
-                      );
-                    },
-                  },
-                  {
-                    field: "Plan Type",
-                    flex: 1,
-                    valueGetter(params) {
-                      return params.row._source.seatool.PLAN_TYPE;
-                    },
-                  },
-                  {
-                    field: "Status",
-                    flex: 1,
-                    valueGetter(params) {
-                      let status = "Unknown"; // not sure what status to use or even what "record.SPW_STATUS[0].SPW_STATUS_DESC" is
-                      if (
-                        params.row._source.seatool.SPW_STATUS &&
-                        params.row._source.seatool.SPW_STATUS[0]
-                      ) {
-                        status =
-                          params.row._source.seatool.SPW_STATUS[0]
-                            .SPW_STATUS_DESC;
-                      }
-                      return status;
-                    },
-                  },
-                  {
-                    field: "Submission Date",
-                    flex: 1,
-                    valueGetter(params) {
-                      return format(
-                        params.row._source.seatool.SUBMISSION_DATE,
-                        "MM/dd/yyyy"
-                      );
-                    },
-                  },
-                  {
-                    field: "Region",
-                    flex: 1,
-                    valueGetter(params) {
-                      return params.row._source.seatool.REGION[0].REGION_NAME;
-                    },
-                  },
-                  {
-                    field: "Type",
-                    flex: 1,
-                    valueGetter(params) {
-                      return params.row._source.programType;
-                    },
-                  },
-                  {
-                    field: "Memo",
-                    flex: 1,
-                    valueGetter(params) {
-                      return params.row._source.seatool?.STATE_PLAN
-                        ?.SUMMARY_MEMO;
-                    },
-                  },
-                ]}
-                rows={(searchData as SearchData[]) || []}
-                getRowId={(row) => row._id}
-                slots={{
-                  toolbar: GridToolbar,
-                }}
-                onRowSelectionModelChange={(newRowSelectionModel) => {
-                  setRowSelectionModel(newRowSelectionModel.toString());
-                }}
-                rowSelectionModel={rowSelectionModel}
-                initialState={{
-                  ...{ searchData },
-                  pagination: { paginationModel: { pageSize: 10 } },
-                }}
-                pageSizeOptions={[5, 10, 25]}
-              />
-            </UI.TabPanel>
-            <UI.TabPanel id="tab-panel--summary" tabLabel="SPAs">
-              <SearchForm
-                handleSearch={handleSearch}
-                setSearchText={setSearchText}
-                searchText={searchText}
-              />
-              <DataGrid
-                columns={[
-                  {
-                    field: "Transmittal ID Number (TIN)",
-                    hideable: false,
-                    flex: 1,
-                    renderCell(params) {
-                      return (
-                        <Link
-                          className="cursor-pointer text-blue-600"
-                          to={`/record?region=${encodeURIComponent(
-                            params.row._source.seatool.STATE_PLAN.STATE_CODE
-                          )}&id=${encodeURIComponent(params.row._id)}`}
-                        >
-                          {params.row._id}
-                        </Link>
-                      );
-                    },
-                  },
-                  {
-                    field: "Plan Type",
-                    flex: 1,
-                    valueGetter(params) {
-                      return params.row._source.seatool.PLAN_TYPE;
-                    },
-                  },
-                  {
-                    field: "Status",
-                    flex: 1,
-                    valueGetter(params) {
-                      let status = "Unknown"; // not sure what status to use or even what "record.SPW_STATUS[0].SPW_STATUS_DESC" is
-                      if (
-                        params.row._source.seatool.SPW_STATUS &&
-                        params.row._source.seatool.SPW_STATUS[0]
-                      ) {
-                        status =
-                          params.row._source.seatool.SPW_STATUS[0]
-                            .SPW_STATUS_DESC;
-                      }
-                      return status;
-                    },
-                  },
-                  {
-                    field: "Submission Date",
-                    flex: 1,
-                    valueGetter(params) {
-                      return format(
-                        params.row._source.seatool.SUBMISSION_DATE,
-                        "MM/dd/yyyy"
-                      );
-                    },
-                  },
-                  {
-                    field: "Region",
-                    flex: 1,
-                    valueGetter(params) {
-                      return params.row._source.seatool.REGION[0].REGION_NAME;
-                    },
-                  },
-                  {
-                    field: "Type",
-                    flex: 1,
-                    valueGetter(params) {
-                      return params.row._source.programType;
-                    },
-                  },
-                  {
-                    field: "Memo",
-                    flex: 1,
-                    valueGetter(params) {
-                      return params.row._source.seatool?.STATE_PLAN
-                        ?.SUMMARY_MEMO;
-                    },
-                  },
-                ]}
-                rows={(searchData as SearchData[]) || []}
-                getRowId={(row) => row._id}
-                slots={{
-                  toolbar: GridToolbar,
-                }}
-                onRowSelectionModelChange={(newRowSelectionModel) => {
-                  setRowSelectionModel(newRowSelectionModel.toString());
-                }}
-                rowSelectionModel={rowSelectionModel}
-                initialState={{
-                  ...{ searchData },
-                  pagination: { paginationModel: { pageSize: 10 } },
-                }}
-                pageSizeOptions={[5, 10, 25]}
-              />
-            </UI.TabPanel>
-          </UI.Tabs>
-        </div>
-      )}
+      <div
+        style={{
+          width: "100%",
+          alignItems: "center",
+          justifyContent: "center",
+        }}
+      >
+        <UI.Tabs>
+          <UI.TabPanel id="tab-panel--spas" tabLabel="SPAs">
+            <SpasList selectedState={selectedState} />
+          </UI.TabPanel>
+          <UI.TabPanel id="tab-panel--waivers" tabLabel="Waivers">
+            <WaiversList selectedState={selectedState} />
+          </UI.TabPanel>
+        </UI.Tabs>
+      </div>
     </div>
   );
 };
